+++ source/ChanSort.Api/Model/Enums.cs
@@ -73,7 +73,7 @@ namespace ChanSort.Api
   #endregion
 
   [Flags]
-  public enum Favorites : byte { A = 0x01, B = 0x02, C = 0x04, D = 0x08, E = 0x10 }
+  public enum Favorites : byte { A = 0x01, B = 0x02, C = 0x04, D = 0x08, E = 0x10, F=0x20, G=0x40, H=0x80 }
 
   public enum UnsortedChannelMode
   {