+++ source/ChanSort/MainForm.Designer.cs
@@ -41,15 +41,19 @@
       this.colOutName = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colOutFav = new DevExpress.XtraGrid.Columns.GridColumn();
       this.repositoryItemCheckedComboBoxEdit1 = new DevExpress.XtraEditors.Repository.RepositoryItemCheckedComboBoxEdit();
-      this.colUid1 = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colOutLock = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colOutSkip = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colOutHide = new DevExpress.XtraGrid.Columns.GridColumn();
+      this.colOutDeleted = new DevExpress.XtraGrid.Columns.GridColumn();
+      this.colUid1 = new DevExpress.XtraGrid.Columns.GridColumn();
       this.repositoryItemTextEdit1 = new DevExpress.XtraEditors.Repository.RepositoryItemTextEdit();
       this.lblHotkeyLeft = new DevExpress.XtraEditors.LabelControl();
       this.pnlEditControls = new DevExpress.XtraEditors.PanelControl();
-      this.btnToggleLock = new DevExpress.XtraEditors.SimpleButton();
+      this.btnToggleFavH = new DevExpress.XtraEditors.SimpleButton();
       this.globalImageCollection1 = new ChanSort.Ui.GlobalImageCollection(this.components);
+      this.btnToggleFavG = new DevExpress.XtraEditors.SimpleButton();
+      this.btnToggleFavF = new DevExpress.XtraEditors.SimpleButton();
+      this.btnToggleLock = new DevExpress.XtraEditors.SimpleButton();
       this.btnToggleFavE = new DevExpress.XtraEditors.SimpleButton();
       this.btnToggleFavD = new DevExpress.XtraEditors.SimpleButton();
       this.btnToggleFavC = new DevExpress.XtraEditors.SimpleButton();
@@ -75,6 +79,7 @@
       this.colLock = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colSkip = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colHidden = new DevExpress.XtraGrid.Columns.GridColumn();
+      this.colDeleted = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colEncrypted = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colServiceType = new DevExpress.XtraGrid.Columns.GridColumn();
       this.colServiceTypeName = new DevExpress.XtraGrid.Columns.GridColumn();
@@ -126,18 +131,9 @@
       this.miSort = new DevExpress.XtraBars.BarButtonItem();
       this.miRenum = new DevExpress.XtraBars.BarButtonItem();
       this.miRenumFavByPrNr = new DevExpress.XtraBars.BarButtonItem();
+      this.miCopyCsv = new DevExpress.XtraBars.BarButtonItem();
       this.mnuFavSet = new DevExpress.XtraBars.BarSubItem();
-      this.miFavASet = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavBSet = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavCSet = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavDSet = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavESet = new DevExpress.XtraBars.BarButtonItem();
       this.mnuFavUnset = new DevExpress.XtraBars.BarSubItem();
-      this.miFavAUnset = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavBUnset = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavCUnset = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavDUnset = new DevExpress.XtraBars.BarButtonItem();
-      this.miFavEUnset = new DevExpress.XtraBars.BarButtonItem();
       this.miLockOn = new DevExpress.XtraBars.BarButtonItem();
       this.miLockOff = new DevExpress.XtraBars.BarButtonItem();
       this.miSkipOn = new DevExpress.XtraBars.BarButtonItem();
@@ -208,8 +204,6 @@
       this.pageProgNr = new DevExpress.XtraTab.XtraTabPage();
       this.popupInputSource = new DevExpress.XtraBars.PopupMenu(this.components);
       this.popupFavList = new DevExpress.XtraBars.PopupMenu(this.components);
-      this.colOutDeleted = new DevExpress.XtraGrid.Columns.GridColumn();
-      this.colDeleted = new DevExpress.XtraGrid.Columns.GridColumn();
       ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl1)).BeginInit();
       this.splitContainerControl1.SuspendLayout();
       ((System.ComponentModel.ISupportInitialize)(this.grpOutputList)).BeginInit();
@@ -395,13 +389,6 @@
       this.repositoryItemCheckedComboBoxEdit1.ShowPopupCloseButton = false;
       this.repositoryItemCheckedComboBoxEdit1.TextEditStyle = DevExpress.XtraEditors.Controls.TextEditStyles.Standard;
       // 
-      // colUid1
-      // 
-      resources.ApplyResources(this.colUid1, "colUid1");
-      this.colUid1.FieldName = "Uid";
-      this.colUid1.Name = "colUid1";
-      this.colUid1.OptionsColumn.AllowEdit = false;
-      // 
       // colOutLock
       // 
       resources.ApplyResources(this.colOutLock, "colOutLock");
@@ -420,6 +407,19 @@
       this.colOutHide.FieldName = "Hidden";
       this.colOutHide.Name = "colOutHide";
       // 
+      // colOutDeleted
+      // 
+      resources.ApplyResources(this.colOutDeleted, "colOutDeleted");
+      this.colOutDeleted.FieldName = "IsDeleted";
+      this.colOutDeleted.Name = "colOutDeleted";
+      // 
+      // colUid1
+      // 
+      resources.ApplyResources(this.colUid1, "colUid1");
+      this.colUid1.FieldName = "Uid";
+      this.colUid1.Name = "colUid1";
+      this.colUid1.OptionsColumn.AllowEdit = false;
+      // 
       // repositoryItemTextEdit1
       // 
       resources.ApplyResources(this.repositoryItemTextEdit1, "repositoryItemTextEdit1");
@@ -437,6 +437,9 @@
       // 
       // pnlEditControls
       // 
+      this.pnlEditControls.Controls.Add(this.btnToggleFavH);
+      this.pnlEditControls.Controls.Add(this.btnToggleFavG);
+      this.pnlEditControls.Controls.Add(this.btnToggleFavF);
       this.pnlEditControls.Controls.Add(this.btnToggleLock);
       this.pnlEditControls.Controls.Add(this.btnToggleFavE);
       this.pnlEditControls.Controls.Add(this.btnToggleFavD);
@@ -450,6 +453,28 @@
       this.pnlEditControls.Controls.Add(this.btnRemoveLeft);
       resources.ApplyResources(this.pnlEditControls, "pnlEditControls");
       this.pnlEditControls.Name = "pnlEditControls";
+      // 
+      // btnToggleFavH
+      // 
+      resources.ApplyResources(this.btnToggleFavH, "btnToggleFavH");
+      this.btnToggleFavH.ImageOptions.ImageList = this.globalImageCollection1;
+      this.btnToggleFavH.Name = "btnToggleFavH";
+      this.btnToggleFavH.Tag = "";
+      this.btnToggleFavH.Click += new System.EventHandler(this.btnToggleFav_Click);
+      // 
+      // btnToggleFavG
+      // 
+      resources.ApplyResources(this.btnToggleFavG, "btnToggleFavG");
+      this.btnToggleFavG.ImageOptions.ImageList = this.globalImageCollection1;
+      this.btnToggleFavG.Name = "btnToggleFavG";
+      this.btnToggleFavG.Click += new System.EventHandler(this.btnToggleFav_Click);
+      // 
+      // btnToggleFavF
+      // 
+      resources.ApplyResources(this.btnToggleFavF, "btnToggleFavF");
+      this.btnToggleFavF.ImageOptions.ImageList = this.globalImageCollection1;
+      this.btnToggleFavF.Name = "btnToggleFavF";
+      this.btnToggleFavF.Click += new System.EventHandler(this.btnToggleFav_Click);
       // 
       // btnToggleLock
       // 
@@ -733,6 +758,12 @@
       this.colHidden.FieldName = "Hidden";
       this.colHidden.Name = "colHidden";
       // 
+      // colDeleted
+      // 
+      resources.ApplyResources(this.colDeleted, "colDeleted");
+      this.colDeleted.FieldName = "IsDeleted";
+      this.colDeleted.Name = "colDeleted";
+      // 
       // colEncrypted
       // 
       resources.ApplyResources(this.colEncrypted, "colEncrypted");
@@ -1006,16 +1037,6 @@
             this.miGotoLeftList,
             this.miRightListFilter,
             this.miGotoRightList,
-            this.miFavASet,
-            this.miFavBSet,
-            this.miFavCSet,
-            this.miFavDSet,
-            this.miFavESet,
-            this.miFavAUnset,
-            this.miFavBUnset,
-            this.miFavCUnset,
-            this.miFavDUnset,
-            this.miFavEUnset,
             this.miSelectFavList0,
             this.miSelectFavListA,
             this.miSelectFavListB,
@@ -1032,9 +1053,10 @@
             this.miRomanian,
             this.miExplorerIntegration,
             this.miCheckUpdates,
-            this.miUtf8Charset});
+            this.miUtf8Charset,
+            this.miCopyCsv});
       this.barManager1.MainMenu = this.bar1;
-      this.barManager1.MaxItemId = 100;
+      this.barManager1.MaxItemId = 101;
       this.barManager1.ShowFullMenus = true;
       // 
       // bar1
@@ -1229,17 +1251,18 @@
             new DevExpress.XtraBars.LinkPersistInfo(this.miAddChannel),
             new DevExpress.XtraBars.LinkPersistInfo(this.miRemove),
             new DevExpress.XtraBars.LinkPersistInfo(this.miRenameChannel),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miSort),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miRenum),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miRenumFavByPrNr),
-            new DevExpress.XtraBars.LinkPersistInfo(DevExpress.XtraBars.BarLinkUserDefines.PaintStyle, this.mnuFavSet, "", true, true, true, 0, null, DevExpress.XtraBars.BarItemPaintStyle.CaptionGlyph),
+            new DevExpress.XtraBars.LinkPersistInfo(DevExpress.XtraBars.BarLinkUserDefines.PaintStyle, this.mnuFavSet, DevExpress.XtraBars.BarItemPaintStyle.CaptionGlyph),
             new DevExpress.XtraBars.LinkPersistInfo(this.mnuFavUnset),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miLockOn, true),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miLockOn),
             new DevExpress.XtraBars.LinkPersistInfo(this.miLockOff),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miSkipOn, true),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miSkipOn),
             new DevExpress.XtraBars.LinkPersistInfo(this.miSkipOff),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOn, true),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOff)});
+            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOn),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOff),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miRenum, true),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miSort),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miRenumFavByPrNr),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miCopyCsv)});
       this.miEdit.Name = "miEdit";
       // 
       // miAddChannel
@@ -1297,138 +1320,31 @@
       this.miRenumFavByPrNr.Name = "miRenumFavByPrNr";
       this.miRenumFavByPrNr.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miRenumFavByPrNr_ItemClick);
       // 
+      // miCopyCsv
+      // 
+      resources.ApplyResources(this.miCopyCsv, "miCopyCsv");
+      this.miCopyCsv.Id = 100;
+      this.miCopyCsv.ImageOptions.ImageIndex = ((int)(resources.GetObject("miCopyCsv.ImageOptions.ImageIndex")));
+      this.miCopyCsv.ItemShortcut = new DevExpress.XtraBars.BarShortcut(((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.Shift) 
+                | System.Windows.Forms.Keys.C));
+      this.miCopyCsv.Name = "miCopyCsv";
+      this.miCopyCsv.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miCopyCsv_ItemClick);
+      // 
       // mnuFavSet
       // 
       resources.ApplyResources(this.mnuFavSet, "mnuFavSet");
       this.mnuFavSet.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
       this.mnuFavSet.Id = 37;
       this.mnuFavSet.ImageOptions.ImageIndex = ((int)(resources.GetObject("mnuFavSet.ImageOptions.ImageIndex")));
-      this.mnuFavSet.LinksPersistInfo.AddRange(new DevExpress.XtraBars.LinkPersistInfo[] {
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavASet),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavBSet),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavCSet),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavDSet),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavESet)});
       this.mnuFavSet.Name = "mnuFavSet";
       this.mnuFavSet.PaintStyle = DevExpress.XtraBars.BarItemPaintStyle.CaptionInMenu;
-      // 
-      // miFavASet
-      // 
-      resources.ApplyResources(this.miFavASet, "miFavASet");
-      this.miFavASet.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavASet.Id = 73;
-      this.miFavASet.ItemShortcut = new DevExpress.XtraBars.BarShortcut((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.D1));
-      this.miFavASet.Name = "miFavASet";
-      this.miFavASet.Tag = "A";
-      this.miFavASet.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavSet_ItemClick);
-      // 
-      // miFavBSet
-      // 
-      resources.ApplyResources(this.miFavBSet, "miFavBSet");
-      this.miFavBSet.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavBSet.Id = 74;
-      this.miFavBSet.ItemShortcut = new DevExpress.XtraBars.BarShortcut((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.D2));
-      this.miFavBSet.Name = "miFavBSet";
-      this.miFavBSet.Tag = "B";
-      this.miFavBSet.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavSet_ItemClick);
-      // 
-      // miFavCSet
-      // 
-      resources.ApplyResources(this.miFavCSet, "miFavCSet");
-      this.miFavCSet.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavCSet.Id = 75;
-      this.miFavCSet.ItemShortcut = new DevExpress.XtraBars.BarShortcut((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.D3));
-      this.miFavCSet.Name = "miFavCSet";
-      this.miFavCSet.Tag = "C";
-      this.miFavCSet.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavSet_ItemClick);
-      // 
-      // miFavDSet
-      // 
-      resources.ApplyResources(this.miFavDSet, "miFavDSet");
-      this.miFavDSet.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavDSet.Id = 76;
-      this.miFavDSet.ItemShortcut = new DevExpress.XtraBars.BarShortcut((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.D4));
-      this.miFavDSet.Name = "miFavDSet";
-      this.miFavDSet.Tag = "D";
-      this.miFavDSet.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavSet_ItemClick);
-      // 
-      // miFavESet
-      // 
-      resources.ApplyResources(this.miFavESet, "miFavESet");
-      this.miFavESet.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavESet.Id = 77;
-      this.miFavESet.ItemShortcut = new DevExpress.XtraBars.BarShortcut((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.D5));
-      this.miFavESet.Name = "miFavESet";
-      this.miFavESet.Tag = "E";
-      this.miFavESet.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavSet_ItemClick);
       // 
       // mnuFavUnset
       // 
       resources.ApplyResources(this.mnuFavUnset, "mnuFavUnset");
       this.mnuFavUnset.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
       this.mnuFavUnset.Id = 38;
-      this.mnuFavUnset.LinksPersistInfo.AddRange(new DevExpress.XtraBars.LinkPersistInfo[] {
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavAUnset),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavBUnset),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavCUnset),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavDUnset),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miFavEUnset)});
       this.mnuFavUnset.Name = "mnuFavUnset";
-      // 
-      // miFavAUnset
-      // 
-      resources.ApplyResources(this.miFavAUnset, "miFavAUnset");
-      this.miFavAUnset.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavAUnset.Id = 78;
-      this.miFavAUnset.ItemShortcut = new DevExpress.XtraBars.BarShortcut(((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.Shift) 
-                | System.Windows.Forms.Keys.D1));
-      this.miFavAUnset.Name = "miFavAUnset";
-      this.miFavAUnset.Tag = "A";
-      this.miFavAUnset.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavUnset_ItemClick);
-      // 
-      // miFavBUnset
-      // 
-      resources.ApplyResources(this.miFavBUnset, "miFavBUnset");
-      this.miFavBUnset.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavBUnset.Id = 79;
-      this.miFavBUnset.ItemShortcut = new DevExpress.XtraBars.BarShortcut(((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.Shift) 
-                | System.Windows.Forms.Keys.D2));
-      this.miFavBUnset.Name = "miFavBUnset";
-      this.miFavBUnset.Tag = "B";
-      this.miFavBUnset.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavUnset_ItemClick);
-      // 
-      // miFavCUnset
-      // 
-      resources.ApplyResources(this.miFavCUnset, "miFavCUnset");
-      this.miFavCUnset.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavCUnset.Id = 80;
-      this.miFavCUnset.ItemShortcut = new DevExpress.XtraBars.BarShortcut(((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.Shift) 
-                | System.Windows.Forms.Keys.D3));
-      this.miFavCUnset.Name = "miFavCUnset";
-      this.miFavCUnset.Tag = "C";
-      this.miFavCUnset.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavUnset_ItemClick);
-      // 
-      // miFavDUnset
-      // 
-      resources.ApplyResources(this.miFavDUnset, "miFavDUnset");
-      this.miFavDUnset.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavDUnset.Id = 81;
-      this.miFavDUnset.ItemShortcut = new DevExpress.XtraBars.BarShortcut(((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.Shift) 
-                | System.Windows.Forms.Keys.D4));
-      this.miFavDUnset.Name = "miFavDUnset";
-      this.miFavDUnset.Tag = "D";
-      this.miFavDUnset.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavUnset_ItemClick);
-      // 
-      // miFavEUnset
-      // 
-      resources.ApplyResources(this.miFavEUnset, "miFavEUnset");
-      this.miFavEUnset.CategoryGuid = new System.Guid("d7eec464-59c9-4f45-88aa-602e64c81cc0");
-      this.miFavEUnset.Id = 82;
-      this.miFavEUnset.ItemShortcut = new DevExpress.XtraBars.BarShortcut(((System.Windows.Forms.Keys.Control | System.Windows.Forms.Keys.Shift) 
-                | System.Windows.Forms.Keys.D5));
-      this.miFavEUnset.Name = "miFavEUnset";
-      this.miFavEUnset.Tag = "E";
-      this.miFavEUnset.ItemClick += new DevExpress.XtraBars.ItemClickEventHandler(this.miFavUnset_ItemClick);
       // 
       // miLockOn
       // 
@@ -2020,16 +1936,17 @@
             new DevExpress.XtraBars.LinkPersistInfo(this.miAddChannel),
             new DevExpress.XtraBars.LinkPersistInfo(this.miRemove),
             new DevExpress.XtraBars.LinkPersistInfo(this.miRenameChannel),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miSort),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miRenum),
-            new DevExpress.XtraBars.LinkPersistInfo(DevExpress.XtraBars.BarLinkUserDefines.PaintStyle, this.mnuFavSet, "", true, true, true, 0, null, DevExpress.XtraBars.BarItemPaintStyle.CaptionGlyph),
+            new DevExpress.XtraBars.LinkPersistInfo(DevExpress.XtraBars.BarLinkUserDefines.PaintStyle, this.mnuFavSet, DevExpress.XtraBars.BarItemPaintStyle.CaptionGlyph),
             new DevExpress.XtraBars.LinkPersistInfo(this.mnuFavUnset),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miLockOn, true),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miLockOn),
             new DevExpress.XtraBars.LinkPersistInfo(this.miLockOff),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miSkipOn, true),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miSkipOn),
             new DevExpress.XtraBars.LinkPersistInfo(this.miSkipOff),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOn, true),
-            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOff)});
+            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOn),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miHideOff),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miRenum, true),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miSort),
+            new DevExpress.XtraBars.LinkPersistInfo(this.miCopyCsv)});
       this.popupContext.Manager = this.barManager1;
       this.popupContext.Name = "popupContext";
       // 
@@ -2076,18 +1993,6 @@
       this.popupFavList.MenuCaption = "Select program or favorite list";
       this.popupFavList.Name = "popupFavList";
       this.popupFavList.ShowCaption = true;
-      // 
-      // colOutDeleted
-      // 
-      resources.ApplyResources(this.colOutDeleted, "colOutDeleted");
-      this.colOutDeleted.FieldName = "IsDeleted";
-      this.colOutDeleted.Name = "colOutDeleted";
-      // 
-      // colDeleted
-      // 
-      resources.ApplyResources(this.colDeleted, "colDeleted");
-      this.colDeleted.FieldName = "IsDeleted";
-      this.colDeleted.Name = "colDeleted";
       // 
       // MainForm
       // 
@@ -2297,16 +2202,6 @@
     private DevExpress.XtraBars.BarButtonItem miGotoLeftList;
     private DevExpress.XtraBars.BarButtonItem miRightListFilter;
     private DevExpress.XtraBars.BarButtonItem miGotoRightList;
-    private DevExpress.XtraBars.BarButtonItem miFavASet;
-    private DevExpress.XtraBars.BarButtonItem miFavBSet;
-    private DevExpress.XtraBars.BarButtonItem miFavCSet;
-    private DevExpress.XtraBars.BarButtonItem miFavDSet;
-    private DevExpress.XtraBars.BarButtonItem miFavESet;
-    private DevExpress.XtraBars.BarButtonItem miFavAUnset;
-    private DevExpress.XtraBars.BarButtonItem miFavBUnset;
-    private DevExpress.XtraBars.BarButtonItem miFavCUnset;
-    private DevExpress.XtraBars.BarButtonItem miFavDUnset;
-    private DevExpress.XtraBars.BarButtonItem miFavEUnset;
     private DevExpress.XtraBars.PopupMenu popupInputSource;
     private DevExpress.XtraBars.PopupMenu popupFavList;
     private DevExpress.XtraBars.BarButtonItem miSelectFavList0;
@@ -2333,6 +2228,10 @@
     private DevExpress.XtraBars.BarButtonItem miUtf8Charset;
     private DevExpress.XtraGrid.Columns.GridColumn colOutDeleted;
     private DevExpress.XtraGrid.Columns.GridColumn colDeleted;
+    private DevExpress.XtraEditors.SimpleButton btnToggleFavH;
+    private DevExpress.XtraEditors.SimpleButton btnToggleFavG;
+    private DevExpress.XtraEditors.SimpleButton btnToggleFavF;
+    private DevExpress.XtraBars.BarButtonItem miCopyCsv;
   }
 }